backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.15.1
elasticsearch==8.15.1
elasticsearch-dsl==8.15.4
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
